get_charts: move to a new class
